Use float rangeInKm in test as well as regular code path

+++ bobo-contrib/src/test/java/com/browseengine/bobo/geosearch/solo/search/impl/GeoOnlySearcherTest.java
@@ -387,7 +387,7 @@ public class GeoOnlySearcherTest {
     
     private CartesianCoordinateUUID buildMaxCoordinate(double centroidLong, double centroidLat, float rangeInMiles, 
             byte[] uuid) {
-        double rangeInkm = Conversions.mi2km(rangeInMiles);
+        float rangeInkm = Conversions.mi2km(rangeInMiles);
         int rangeInUnits = Conversions.radiusMetersToIntegerUnits(rangeInkm * 1000);
         CartesianCoordinateUUID centroidCoordinate = geoConverter.toCartesianCoordinate(
                 centroidLat, centroidLong, GeoOnlySearcher.EMPTY_UUID);
